pyproject.toml
[tool.pdm]
distribution = false
[tool.pdm.scripts]
_.env_file = ".env"
socketio = 'uvicorn fliji_sockets.main:asgi_app --proxy-headers --host 0.0.0.0 --port 8097 --reload'
socketio-prod = 'uvicorn fliji_sockets.main:asgi_app --proxy-headers --host 0.0.0.0 --port 80'
docs = "sphinx-build -b html docs/source docs/_build"
test = "pytest -s tests/test_main.py"

[project]
name = "fliji-sockets"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "", email = "none" },
]
dependencies = [
    "python-socketio[asyncio]>=5.11.4",
    "uvicorn[standard]>=0.27.1",
    "pymongo>=4.6.1",
    "httpx>=0.26.0",
    "nats-py>=2.7.2",
    "sentry-sdk>=2.4.0",
    "pydantic>=2.9.1",
    # python-socketio dependencies
    "redis>=5.0.8",
    "websocket-client>=1.8.0",
    "PyJWT>=2.9.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "Proprietary" }


[dependency-groups]
dev = [
    "sphinx>=7.2.6",
    "autodoc-pydantic>=2.1.0",
    "pytest>=8.3.3",
    "starlette>=0.38.5",
    "ruff>=0.9.3",
]
